requirements.txt
elasticsearch==6.1.1
lxml==3.3.3
Mako==0.9.1
MarkupSafe==0.18
matplotlib==2.1.2
MySQL-python==1.2.5
numpy==1.10.0
oauthlib==0.6.1
pbr==3.1.1
pexpect==3.1
Pillow==2.3.0
PyAudio==0.2.11
pycrypto==2.6.1
PyDejavu==0.1.3
pydub==0.20.0
pyparsing==2.2.0
pyserial==2.6
pytz==2017.3
pyxdg==0.25
requests==2.2.1
scipy==1.0.0
six==1.11.0
urllib3==1.22
XlsxWriter==1.0.2
zope.interface==4.0.5
